Fix antismash_to_gff, it was producing an incorrect output
Minor refactor:
- rename files to make it more consistent
- removed some __init__.py (we need to remove most of them)
- Added some FIXME comments on some scripts that we need to fix
  mostly documentation

Unit test for changes in temp cwl_tests folder
