+++ cwl_tests/README.md
@@ -1,0 +1,45 @@
+# CWL Tests TMP folder
+
+This folder container some CWL unit tests. We are actively working on this on a different branch (`fix-dockers`) and to avoid issues with the code migration in the future.
+
+## How to run them
+
+It requires `cwltest`, `cwltool`, `tabix` and `biopython`.
+
+```bash
+./run_tests.sh
+```
+
+## antiSMASH notes
+
+This could be useful to document the GFF file build process:
+
+Example for one antiSMASH predicition in the config "ERZ129262-117", feature "ERZ129262-117_41".
+
+GFF entry:
+```
+ERZ1292626.117-Hain-H43-01um-R3-117-length-98827-cov-25 antiSMASH       CDS     40601   41567   .       +       .       ID=ERZ1292626.117-Hain-H43-01um-R3-117-length-98827-cov-25;as_notes=smCOG%3A%20SMCOG1066%3Aalpha/beta_hydrolase_domain-containing_protein%20%28Score%3A%20279.6%3B%20E-value%3A%204.8e-85%29%3B,;as_gene_clusters=nrps;as_type=biosynthetic-additional
+```
+
+the GFF entry was generated by the combination of:
+
+EMBL file entry
+```
+FT   CDS             40601..41566
+FT                   /locus_tag="ERZ129262-117_41"
+FT                   /note="smCOG:
+FT                   SMCOG1066:alpha/beta_hydrolase_domain-containing_protein
+FT                   (Score: 279.6; E-value: 4.8e-85);"
+FT                   /note="smCOG tree PNG image: smcogs/ctg46_41.png"
+FT                   /translation="MALHPDLQAFLELANAQASSRPRMSQMTAAQARAAYDNATLALDL
+FT                   PGDDLPTEDLDLPTRDGARIRARLYRAAPEYTAAPCLLFLHGGGYVLGGLDSHDSLCRS
+FT                   LALGTPCNVLAIDYRLAPEHRFPTAFLDAQDALAWLLEHGGRHGLDAQRLAIGGDSVGG
+FT                   TLATALCIAARDAGMPQPLLQLLLYPCTAAQQDSDSHARLATGHLLEDDTLAWMFGQYL
+FT                   RSDADRQDWRFAPLQASDLGRLAPTLMVLAEYDPLIDEGKRYAERLQAAGVAVDLQIYA
+FT                   GMVHDFARLGALTEETEVLRGRICKALAHAFQAARIPALA"
+```
+
+and this genecluster.txt row:
+```tsv
+ERZ129262-117	ERZ129262-117	nrps	ERZ129262-117_31;ERZ129262-117_32;ERZ129262-117_33;ERZ129262-117_34;ERZ129262-117_35;ERZ129262-117_36;ERZ129262-117_37;ERZ129262-117_38;ERZ129262-117_39;ERZ129262-117_40;ERZ129262-117_41;ERZ129262-117_42;ERZ129262-117_43;ERZ129262-117_44;ERZ129262-117_45;ERZ129262-117_46;ERZ129262-117_47;ERZ129262-117_48;ERZ129262-117_49;ERZ129262-117_50;ERZ129262-117_51;ERZ129262-117_52;ERZ129262-117_53;ERZ129262-117_54;ERZ129262-117_55;ERZ129262-117_56;ERZ129262-117_57;ERZ129262-117_58;ERZ129262-117_59;ERZ129262-117_60;ERZ129262-117_61;ERZ129262-117_62;ERZ129262-117_63	ERZ129262-117_31;ERZ129262-117_32;ERZ129262-117_33;ERZ129262-117_34;ERZ129262-117_35;ERZ129262-117_36;ERZ129262-117_37;ERZ129262-117_38;ERZ129262-117_39;ERZ129262-117_40;ERZ129262-117_41;ERZ129262-117_42;ERZ129262-117_43;ERZ129262-117_44;ERZ129262-117_45;ERZ129262-117_46;ERZ129262-117_47;ERZ129262-117_48;ERZ129262-117_49;ERZ129262-117_50;ERZ129262-117_51;ERZ129262-117_52;ERZ129262-117_53;ERZ129262-117_54;ERZ129262-117_55;ERZ129262-117_56;ERZ129262-117_57;ERZ129262-117_58;ERZ129262-117_59;ERZ129262-117_60;ERZ129262-117_61;ERZ129262-117_62;ERZ129262-117_63	Non-ribosomal peptide synthetase cluster
+```
