+++ cwl_tests/tools/antismash_gff/antismash_gff.yml
@@ -1,0 +1,7 @@
+antismash_geneclus:
+  class: File
+  path: antismash_geneclusters.txt
+antismash_embl:
+  class: File
+  path: antismash_final.embl
+output_name: test_antismash.gff
