+++ tools/Assembly/antismash/chunking_antismash_with_conditionals/post-processing/gff_antismash/antismash_to_gff.yml
@@ -1,0 +1,7 @@
+antismash_geneclus:
+  class: File
+  path: ../antismash/expected-output/assembly-SRR1620013_small_FASTA_FASTA_antismash_geneclusters.txt
+antismash_embl:
+  class: File
+  path: ../antismash/expected-output/assembly-SRR1620013_small_FASTA_FASTA_antismash_final.embl.gz
+output_name: assembly-SRR1620013_small_FASTA_FASTA.antismash.gff
