--- tools/Assembly/antismash/old_run_antismash/wf_antismash.yml
@@ -1,10 +1,0 @@
-filtered_fasta:
-  class: File
-  path: ../../../input_examples/assembly/assembly-SRR1620013_small_FASTA_FASTA.fasta
-
-clusters_glossary:
-  class: File
-  path: ref-dbs/antismash_glossary.tsv
-  checksum: bd50e85246dc3609f7cab6fb4efe4845
-
-final_folder_name: pathways-systems
